There is only 1 inheritor of SnsTopicBase, so let's move it all into 1 class and remove any reundant code
