+++ JustSaying.AwsTools/JustSaying.AwsTools.csproj
@@ -73,7 +73,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="QueueCreation\SqsWriteConfiguration.cs" />
     <Compile Include="SnsTopicByName.cs" />
-    <Compile Include="SnsTopicBase.cs" />
     <Compile Include="SqsNotificationListener.cs" />
     <Compile Include="SqsPublisher.cs" />
     <Compile Include="SqsQueueBase.cs" />
